feat(blob): Moving asset/model-stream to model-blob
