+++ module/rest/src/interceptor/serialize-util.ts
@@ -1,8 +1,8 @@
 import { Readable } from 'node:stream';
 
-import { ErrorCategory, AppError, BlobUtil } from '@travetto/runtime';
+import { ErrorCategory, AppError } from '@travetto/runtime';
 
-import { HeadersAddedⲐ, NodeEntityⲐ } from '../internal/symbol';
+import { HeadersAddedⲐ } from '../internal/symbol';
 import { Renderable } from '../response/renderable';
 import { Request, Response } from '../types';
 
@@ -105,7 +105,7 @@ export class SerializeUtil {
    * Serialize file/blob
    */
   static async serializeBlob(res: Response, output: Blob | File): Promise<void> {
-    const meta = BlobUtil.getBlobMeta(output);
+    const meta = output.meta;
     if (meta) {
       res.statusCode = meta?.range ? 206 : 200;
       for (const [k, v] of Object.entries({
