+++ module/runtime/__index__.ts
@@ -1,7 +1,7 @@
 /// <reference path="./src/trv.d.ts" />
 /// <reference path="./src/global.d.ts" />
 export * from './src/blob';
-export * from './src/bytes';
+export * from './src/binary';
 export * from './src/console';
 export * from './src/context';
 export * from './src/debug';
