+++ module/runtime/src/global.d.ts
@@ -1,4 +1,4 @@
-import 'node:os';
+import { BlobMeta } from './types';
 
 // https://github.com/microsoft/TypeScript/issues/59012
 declare const write: unique symbol;
@@ -8,5 +8,13 @@ declare global {
   }
   interface Function {
     Ⲑid: string;
+  }
+
+  interface Blob {
+    readonly meta?: Readonly<BlobMeta>;
+  }
+
+  interface File {
+    readonly meta?: Readonly<BlobMeta>;
   }
 }
