+++ module/runtime/src/types.ts
@@ -10,7 +10,6 @@ export type ClassInstance<T = Any> = T & {
   constructor: Class<T> & { Ⲑid: string };
 };
 
-
 export type BinaryInput = Blob | Buffer | Readable | ReadableStream;
 
 export type TypedFunction<R = Any, V = unknown> = (this: V, ...args: Any[]) => R;
@@ -45,4 +44,30 @@ export const asConstructable = <Z = unknown>(input: Class | unknown): { construc
 export function classConstruct<T>(cls: Class<T>, args: unknown[] = []): ClassInstance<T> {
   const cons: { new(..._args: Any[]): T } = castTo(cls);
   return castTo(new cons(...args));
+}
+
+/**
+ * Range of bytes, inclusive
+ */
+export type ByteRange = { start: number, end?: number };
+
+export interface BlobMeta {
+  /** Size of blob */
+  size?: number;
+  /** Mime type of the content */
+  contentType?: string;
+  /** Hash of blob contents */
+  hash?: string;
+  /** The original base filename of the file */
+  filename?: string;
+  /** Filenames title, optional for elements like images, audio, videos */
+  title?: string;
+  /** Content encoding */
+  contentEncoding?: string;
+  /** Content language */
+  contentLanguage?: string;
+  /** Cache control */
+  cacheControl?: string;
+  /** Byte range for blob */
+  range?: Required<ByteRange>;
 }
