+++ module/runtime/src/util.ts
@@ -1,8 +1,5 @@
 import crypto from 'node:crypto';
 import timers from 'node:timers/promises';
-import fs from 'node:fs/promises';
-import path from 'node:path';
-import os from 'node:os';
 
 import { castTo } from './types';
 
@@ -58,18 +55,6 @@ export class Util {
   }
 
   /**
-   * Generate a proper sha512 hash from a src value
-   * @param src The seed value to build the hash from
-   * @param len The optional length of the hash to generate
-   */
-  static hash(src: string, len: number = -1): string {
-    const hash = crypto.createHash('sha512');
-    hash.update(src);
-    const ret = hash.digest('hex');
-    return len > 0 ? ret.substring(0, len) : ret;
-  }
-
-  /**
    * Produce a promise that is externally resolvable
    */
   static resolvablePromise<T = void>(): PromiseWithResolvers<T> {
@@ -96,23 +81,6 @@ export class Util {
         yield m;
       }
     }
-  }
-
-  /**
-   * Write file and copy over when ready
-   */
-  static async bufferedFileWrite(file: string, content: string, checkHash = false): Promise<void> {
-    if (checkHash) {
-      const current = await fs.readFile(file, 'utf8').catch(() => '');
-      if (Util.hash(current) === Util.hash(content)) {
-        return;
-      }
-    }
-
-    const temp = path.resolve(os.tmpdir(), `${process.hrtime()[1]}.${path.basename(file)}`);
-    await fs.writeFile(temp, content, 'utf8');
-    await fs.mkdir(path.dirname(file), { recursive: true });
-    await fs.rename(temp, file);
   }
 
   /**
